+++ src/components/Intro/Corner/index.tsx
