+++ src/components/Intro/LeftBlock/Finder/index.tsx
