+++ src/components/Intro/LeftBlock/Watermark/__tests__/index.test.tsx
@@ -1,0 +1,12 @@
+import { render } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import { renderWithTheme } from '../../../../../utils/theme';
+import Watermark from '..';
+
+describe('<Watermark />', () => {
+  it('displays the content properly', () => {
+    const { container } = render(renderWithTheme(<Watermark />));
+
+    expect(container).toMatchSnapshot();
+  });
+});
