--- src/components/Intro/Logo.tsx
@@ -1,164 +1,0 @@
-import { colors } from '../../utils/colors';
-
-const Logo = () => (
-  <svg width="280" height="79" viewBox="0 0 280 79" fill="none">
-    <path
-      d="M36.9724 58.3775H0L10.2971 0H24.8915L16.8646 45.729H39.2426L36.9724 58.3775Z"
-      fill="url(#paint4_linear_105_2)"
-    />
-    <g>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M80.4327 32.6833C81.9914 34.3672 82.7765 36.4265 82.7765 38.8272C82.7765 40.3859 82.5944 41.7853 82.2417 43.0368C81.8777 44.277 81.4226 45.2896 80.8878 46.0746C80.3303 46.8596 79.6591 47.5423 78.8626 48.1339C78.0548 48.7256 77.3153 49.1579 76.644 49.431C75.9728 49.704 75.1991 49.9202 74.323 50.0681C73.4583 50.216 72.8098 50.2956 72.3547 50.3298L70.9326 50.3639L69.7948 50.398L68.2133 50.4322L66.6319 50.398L65.4941 50.3639L64.0719 50.3298C63.6168 50.307 62.9683 50.2274 62.1036 50.0681C61.2276 49.9202 60.4653 49.7154 59.7826 49.431C59.1114 49.1579 58.3718 48.7256 57.564 48.1339C56.7562 47.5423 56.085 46.8596 55.5275 46.0746C54.9814 45.2896 54.5263 44.277 54.1736 43.0368C53.8209 41.7853 53.6388 40.3859 53.6388 38.8272C53.6388 36.4265 54.4239 34.3786 55.9826 32.6833C55.8119 32.6037 55.8005 31.7618 55.9598 30.1575C56.0963 28.5533 56.4377 27.0742 56.9952 25.7203C58.9407 25.9251 61.3414 27.0287 64.2085 29.0084C65.1756 28.7581 66.5067 28.6216 68.202 28.6216C69.9768 28.6216 71.308 28.7467 72.1954 29.0084C73.5039 28.121 74.7554 27.4042 75.95 26.8581C77.156 26.312 78.0321 25.9934 78.5896 25.9137L79.4088 25.7203C79.9663 27.0742 80.3076 28.5533 80.4441 30.1575C80.6148 31.7618 80.6034 32.6037 80.4327 32.6833ZM68.2703 49.0357C71.7745 49.0357 74.4141 48.6147 76.2117 47.7728C77.998 46.9309 78.9082 45.2015 78.9082 42.5733C78.9082 41.0601 78.3393 39.7858 77.1902 38.7732C76.6099 38.2271 75.9159 37.8858 75.1309 37.7607C74.3572 37.6355 73.1626 37.6355 71.5583 37.7607C69.9541 37.8858 68.8619 37.9541 68.2703 37.9541C67.4625 37.9541 66.5978 37.9086 65.5169 37.8289L62.9797 37.6696C62.3654 37.6469 61.7055 37.7265 60.9887 37.8972C60.2719 38.0678 59.6803 38.3637 59.2138 38.7846C58.1216 39.7517 57.5755 41.026 57.5755 42.5847C57.5755 45.2015 58.4629 46.9423 60.2378 47.7842C62.0127 48.6261 64.6409 49.0471 68.1451 49.0471L68.2703 49.0357Z"
-        fill="url(#paint5_linear_105_2)"
-      />
-      <path
-        d="M74.6302 40.92C74.232 40.3625 73.7655 40.0894 73.208 40.0894C72.6278 40.0894 72.1272 40.3625 71.7176 40.92C71.3194 41.4661 71.1146 42.1374 71.1146 42.9452C71.1146 43.7416 71.3194 44.4242 71.7176 44.9703C72.1158 45.5165 72.6164 45.7895 73.208 45.7895C73.7542 45.7895 74.2206 45.5165 74.6302 44.9703C75.0284 44.4242 75.2332 43.753 75.2332 42.9452C75.2332 42.1487 75.0284 41.4661 74.6302 40.92Z"
-        fill="url(#paint6_linear_105_2)"
-      />
-      <path
-        d="M64.5331 41.114C64.1349 40.5565 63.6684 40.2834 63.1109 40.2834C62.5307 40.2834 62.03 40.5565 61.6205 41.114C61.2222 41.6601 61.0175 42.3314 61.0175 43.1392C61.0175 43.9356 61.2222 44.6183 61.6205 45.1644C62.0187 45.7105 62.5193 45.9835 63.1109 45.9835C63.657 45.9835 64.1235 45.7105 64.5331 45.1644C64.9313 44.6183 65.1361 43.947 65.1361 43.1392C65.1361 42.3428 64.9313 41.6601 64.5331 41.114Z"
-        fill="url(#paint7_linear_105_2)"
-      />
-      <path
-        d="M67.8899 61.8442C54.4395 61.8442 43.5038 51.2103 43.5038 38.1517C43.5038 25.0931 54.4395 14.4592 67.8899 14.4592C81.3403 14.4592 92.2759 25.0931 92.2759 38.1517C92.2759 51.2103 81.3403 61.8442 67.8899 61.8442ZM67.8899 19.7253C57.4306 19.7253 48.924 27.9899 48.924 38.1517C48.924 48.3135 57.4306 56.5781 67.8899 56.5781C78.3492 56.5781 86.8557 48.3135 86.8557 38.1517C86.8557 27.9899 78.3492 19.7253 67.8899 19.7253Z"
-        fill="url(#paint8_linear_105_2)"
-      />
-      <path
-        d="M80.8158 56.0983C79.7061 55.0175 79.7083 53.2673 80.8208 52.1892C81.9332 51.1111 83.7346 51.1133 84.8443 52.1941C85.9539 53.2749 95.6002 62.6701 95.6002 62.6701C96.7098 63.7509 96.7076 65.5011 95.5952 66.5792C94.4827 67.6573 92.6813 67.6551 91.5717 66.5743L80.8158 56.0983Z"
-        fill="url(#paint9_linear_105_2)"
-      />
-    </g>
-    <path
-      d="M138.421 43.639C138.421 45.0444 139.124 45.747 140.529 45.747C141.989 45.747 143.205 45.4768 144.178 44.9362L143.773 54.0983C141.178 56.2063 138.259 57.2604 135.016 57.2604C128.908 57.2604 125.584 54.9091 125.043 50.2064C121.8 54.9091 117.422 57.2604 111.908 57.2604C106.449 57.2604 102.908 56.0172 101.287 53.5307C100.206 51.9632 99.6651 50.1794 99.6651 48.1794C99.6651 46.1795 99.8273 44.3417 100.152 42.666L104.692 17.1259H118.476L114.665 38.6931C114.503 39.6661 114.422 40.639 114.422 41.612C114.422 44.2065 115.881 45.5038 118.8 45.5038C121.232 45.5038 123.178 44.666 124.638 42.9903L129.259 17.1259H143.043L138.502 42.666C138.448 42.9903 138.421 43.3147 138.421 43.639Z"
-      fill="url(#paint0_linear_105_2)"
-    />
-    <path
-      d="M177.359 16.0718C182.872 16.0718 186.683 17.5313 188.791 20.4501C190.304 22.4501 191.061 24.9906 191.061 28.0716C191.061 29.3149 190.926 30.6392 190.656 32.0446L189.034 41.5309C188.115 46.666 186.142 50.5848 183.115 53.2875C180.088 55.9361 176.305 57.2604 171.764 57.2604C167.224 57.2604 164.062 56.0442 162.278 53.6118L157.899 78.6655H144.116L154.981 17.1259H167.467L167.061 19.315C169.602 17.1529 173.034 16.0718 177.359 16.0718ZM168.926 45.5849C170.44 45.5849 171.845 45.0984 173.142 44.1255C174.44 43.1525 175.251 41.6931 175.575 39.7471L176.629 33.6662C176.737 33.1256 176.791 32.45 176.791 31.6392C176.791 30.8284 176.44 29.9635 175.737 29.0446C175.034 28.1257 173.818 27.6662 172.088 27.6662C169.278 27.6662 167.44 28.3419 166.575 29.6932L164.143 43.3147C164.899 44.8281 166.494 45.5849 168.926 45.5849Z"
-      fill="url(#paint1_linear_105_2)"
-    />
-    <path
-      d="M212.318 57.2604C201.075 57.2604 195.453 53.3686 195.453 45.5849C195.453 44.5038 195.561 43.3417 195.778 42.0985L197.48 32.0446C198.399 26.8554 200.778 22.8825 204.615 20.1258C208.507 17.3151 213.642 15.9097 220.021 15.9097C230.885 15.9097 236.318 20.0177 236.318 28.2338C236.318 29.423 236.209 30.7203 235.993 32.1256L235.02 37.8823H210.291L209.967 39.7471C209.859 40.3958 209.805 40.9904 209.805 41.5309C209.805 44.5038 211.561 45.9903 215.075 45.9903C216.372 45.9903 217.48 45.5849 218.399 44.7741C219.318 43.9633 219.858 43.0174 220.021 41.9363L233.48 42.0985C231.696 52.2064 224.642 57.2604 212.318 57.2604ZM222.048 32.9364C222.264 32.0175 222.372 31.1797 222.372 30.423C222.372 29.6122 222.021 28.8014 221.318 27.9906C220.615 27.1798 219.48 26.7744 217.912 26.7744C216.399 26.7744 214.994 27.2879 213.696 28.3149C212.453 29.3419 211.588 30.8824 211.102 32.9364H222.048Z"
-      fill="url(#paint2_linear_105_2)"
-    />
-    <path
-      d="M280 25.396C280 27.2338 279.676 29.0176 279.027 30.7473C278.432 32.477 277.838 33.7472 277.243 34.558L276.351 35.6931H264.108C265.46 34.4499 266.135 32.9094 266.135 31.0716C266.135 30.0986 265.811 29.3149 265.162 28.7203C264.568 28.0716 263.73 27.7473 262.649 27.7473C260.811 27.7473 259.217 28.6662 257.865 30.504L253.325 56.1253H239.541L246.433 17.1259H259.244L258.271 22.5582C261.352 18.1799 265.081 15.9907 269.46 15.9907C272.973 15.9907 275.595 16.8286 277.324 18.5042C279.108 20.1799 280 22.4771 280 25.396Z"
-      fill="url(#paint3_linear_105_2)"
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint2_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint3_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint4_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint5_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint6_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint7_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint8_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-      <linearGradient
-        id="paint9_linear_105_2"
-        x1="-12"
-        y1="79"
-        x2="289"
-        y2="79"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor={colors.lightPink} />
-        <stop offset="1" stopColor={colors.lightPurple} />
-      </linearGradient>
-    </defs>
-  </svg>
-);
-
-export default Logo;
