--- src/components/Intro/__tests__/Logo.test.tsx
@@ -1,12 +1,0 @@
-import { render } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import { renderWithTheme } from '../../../utils/theme';
-import Logo from '../Logo';
-
-describe('<Logo />', () => {
-  it('displays the content properly', () => {
-    const { container } = render(renderWithTheme(<Logo />));
-
-    expect(container).toMatchSnapshot();
-  });
-});
