--- src/components/Intro/__tests__/Watermark.test.tsx
@@ -1,12 +1,0 @@
-import { render } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import { renderWithTheme } from '../../../utils/theme';
-import Watermark from '../Watermark';
-
-describe('<Watermark />', () => {
-  it('displays the content properly', () => {
-    const { container } = render(renderWithTheme(<Watermark />));
-
-    expect(container).toMatchSnapshot();
-  });
-});
