+++ src/components/shared/Toast/__tests__/Toast.test.tsx
@@ -1,0 +1,26 @@
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import userEvent from '@testing-library/user-event';
+import Toast from '..';
+
+describe('<Toast />', () => {
+  const props = {
+    isOpen: true,
+    msg: 'Please, choose an available user',
+    onClose: jest.fn(),
+  };
+
+  it('displays the message properly', () => {
+    render(<Toast {...props} />);
+
+    expect(screen.getByText('Please, choose an available user')).toBeInTheDocument();
+  });
+
+  it('calls the correct callback when closing', async () => {
+    render(<Toast {...props} />);
+
+    await userEvent.click(screen.getByRole('button'));
+
+    expect(props.onClose).toHaveBeenCalled();
+  });
+});
