--- src/components/shared/__tests__/Footer.test.tsx
@@ -1,12 +1,0 @@
-import { render } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import { renderWithTheme } from '../../../utils/theme';
-import Footer from '../Footer';
-
-describe('<Footer />', () => {
-  it('displays the content properly', () => {
-    const { container } = render(renderWithTheme(<Footer />));
-
-    expect(container).toMatchSnapshot();
-  });
-});
